changed frame buffer depth from 32 bit to 16 bit because 32 bit is huge overkill

+++ src/CpuFrameBuffer.hpp
@@ -9,10 +9,10 @@
 struct ViewOfCpuFrameBuffer
 {
   uint32_t *image;
-  int32_t *depth;
+  int16_t *depth;
   int w, h;
   
-  void clear(uint32_t argbClearValue = 0xff000000, int32_t depthClearValue = 0) const
+  void clear(uint32_t argbClearValue = 0xff000000, int16_t depthClearValue = 0) const
   {
     std::fill_n(image, w * h, argbClearValue);
     std::fill_n(depth, w * h, depthClearValue);
@@ -24,7 +24,7 @@ struct CpuFrameBuffer
 {
   CpuFrameBuffer(int w, int h)
     : image{std::make_unique<uint32_t[]>(w * h)}
-    , depth{std::make_unique<int32_t[]>(w * h)}
+    , depth{std::make_unique<int16_t[]>(w * h)}
     , w{w}, h{h} {}
   
   void useWith(Function<void(const ViewOfCpuFrameBuffer &)> auto &&f)
@@ -34,6 +34,6 @@ struct CpuFrameBuffer
 
 private:
   const std::unique_ptr<uint32_t[]> image;
-  const std::unique_ptr<int32_t[]> depth;
+  const std::unique_ptr<int16_t[]> depth;
   const int w, h;
 };
